feat: move API key table to dedicated page
- Create new /api-keys page with full API key management functionality
- Add API keys to main navigation with Key2 icon
- Remove API key section from settings page to reduce clutter
- Maintain all existing functionality (create, update, delete, toggle)
- Update settings page description to reflect removed API keys section

This improves the user experience by giving API keys their own dedicated space while keeping settings focused on account and billing management.
